+++ subscription/core/src/main/java/org/occurrent/subscription/OccurrentSubscriptionFilter.java
@@ -23,13 +23,10 @@ import java.util.Objects;
 /**
  * An implementation of {@code SubscriptionFilter} for Occurrent {@link Filter}'s.
  */
-public class OccurrentSubscriptionFilter implements SubscriptionFilter {
+public record OccurrentSubscriptionFilter(Filter filter) implements SubscriptionFilter {
 
-    public final Filter filter;
-
-    public OccurrentSubscriptionFilter(Filter filter) {
+    public OccurrentSubscriptionFilter {
         Objects.requireNonNull(filter, Filter.class.getSimpleName() + " cannot be null");
-        this.filter = filter;
     }
 
     public static OccurrentSubscriptionFilter filter(Filter filter) {